Fixed tests structure
