+++ tests/JsonStreamingParser/Tests/Listener/Consumer.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace JsonStreamingParser\Tests\Listener;
+
+use JsonStreamingParser\Listener\SubsetConsumer;
+
+class Consumer extends SubsetConsumer
+{
+    public $data;
+
+    protected function consume($data)
+    {
+        $this->data = $data;
+        return false;
+    }
+}
