+++ tests/JsonStreamingParser/Tests/Listener/IdealConsumer.php
