+++ tests/JsonStreamingParser/Tests/Listener/SubsetConsumerTest.php
@@ -62,7 +62,7 @@ class SubsetConsumerTest extends \PHPUnit_Framework_TestCase
     public static function differentJsonFiles()
     {
         return array(
-            array(__DIR__ . '/../../example/example.json'),
+            array(__DIR__ . '/../example/example.json'),
             array(__DIR__ . '/data/plain.json'),
             array(__DIR__ . '/data/dateRanges.json'),
             array(__DIR__ . '/data/escapedChars.json'),
